feat(packages/api): add validateSong bump to v1.0.0
