+++ packages/api/package.json
@@ -1,6 +1,6 @@
 {
 	"name": "@song-spotlight/api",
-	"version": "0.1.0",
+	"version": "1.0.0",
 	"description": "Song Spotlight API types and song validation module",
 	"type": "module",
 	"scripts": {
