+++ packages/api/src/handlers/helpers.ts
@@ -1,8 +1,6 @@
 import type { MaybePromise } from "bun";
 import type { Song } from "structs/types";
 
-import type { RenderSongInfo } from "./types";
-
 export interface SongParser {
 	hosts: string[];
 	parse(link: string, host: string, path: string[]): MaybePromise<Song | null>;
@@ -12,5 +10,33 @@ export interface SongService extends SongParser {
 	name: string;
 	types: string[];
 	render(type: string, id: string): MaybePromise<RenderSongInfo | null>;
-	from(type: string, id: string): MaybePromise<string | null>;
+	validate(type: string, id: string): MaybePromise<boolean>;
 }
+
+export interface RenderInfoBase {
+	label: string;
+	sublabel: string;
+	explicit: boolean;
+}
+
+export interface RenderInfoEntry {
+	audio?: {
+		duration: number;
+		previewUrl: string;
+	} | undefined;
+	link: string;
+}
+export type RenderInfoEntryBased = RenderInfoEntry & RenderInfoBase;
+
+interface RenderSongSingle extends RenderInfoBase {
+	form: "single";
+	thumbnailUrl?: string;
+	single: RenderInfoEntry;
+}
+interface RenderSongList extends RenderInfoBase {
+	form: "list";
+	thumbnailUrl?: string;
+	list: RenderInfoEntryBased[];
+}
+
+export type RenderSongInfo = RenderSongSingle | RenderSongList;
