+++ packages/api/src/handlers/index.ts
@@ -1,4 +1,3 @@
 export * from "./core";
 export * from "./finders";
 export * from "./helpers";
-export * from "./types";
